move form into component
