+++ apps/recipes/src/components/recipe-form/recipe-form.types.tsx
@@ -1,0 +1,6 @@
+import { RecipePayload } from '@shared/prisma-payloads/recipe-payload';
+
+export interface RecipeFormProps {
+  recipe?: RecipePayload;
+  mode: 'create' | 'edit';
+}
