+++ apps/recipes/src/shared/measurements.model.ts
@@ -1,10 +1,1 @@
-export enum MeasurementType {
-  GRAMS = 'grams',
-  KGS = 'kgs',
-  TBSP = 'tbsp',
-  TSP = 'tsp',
-  CUP = 'cup',
-  ML = 'ml',
-  LITRE = 'litre',
-  OTHER = 'other',
-}
+export * as MeasurementType from '@generated/enums';
