+++ apps/recipes/src/shared/prisma-payloads/recipe-payload.ts
@@ -1,0 +1,7 @@
+import { RecipeGetPayload } from "@generated/models/Recipe";
+
+export type RecipePayload = RecipeGetPayload<{
+  include: {
+    recipeIngredients: { include: { ingredient: true; measurement: true } };
+  };
+}>;
